Refactor Ws to a separate file
